[REF] mail: move ChannelSelector to discuss folder
Part of task-3265211
