+++ addons/mail/static/src/discuss/core/web/channel_selector.scss
@@ -1,4 +1,4 @@
-.o-mail-ChannelSelector {
+.o-discuss-ChannelSelector {
     input {
         width: unset !important;
     }
