+++ addons/mail/static/src/discuss/core/web/chat_window_patch.js
@@ -1,0 +1,6 @@
+/* @odoo-module */
+
+import { ChannelSelector } from "@mail/discuss/core/web/channel_selector";
+import { ChatWindow } from "@mail/core/common/chat_window";
+
+Object.assign(ChatWindow.components, { ChannelSelector });
