+++ addons/mail/static/src/discuss/core/web/chat_window_patch.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates xml:space="preserve">
+    <t t-inherit="mail.ChatWindow" t-inherit-mode="extension" owl="1">
+        <xpath expr="//t[@name='thread content']" position="after">
+            <div t-if="!thread" class="d-flex align-items-center m-3">
+                <span class="flex-shrink-0 me-2">To :</span>
+                <ChannelSelector category="store.discuss.chats" onValidate="() => chatWindowService.closeNewMessage()" multiple="false" autofocus="true"/>
+            </div>
+        </xpath>
+    </t>
+</templates>
