+++ addons/mail/static/src/discuss/core/web/messaging_menu_patch.js
@@ -1,0 +1,6 @@
+/* @odoo-module */
+
+import { MessagingMenu } from "@mail/core/web/messaging_menu";
+import { ChannelSelector } from "@mail/discuss/core/web/channel_selector";
+
+Object.assign(MessagingMenu.components, { ChannelSelector });
