+++ addons/mail/static/src/discuss/core/web/sidebar_patch.js
@@ -1,0 +1,6 @@
+/* @odoo-module */
+
+import { Sidebar } from "@mail/core/web/sidebar";
+import { ChannelSelector } from "@mail/discuss/core/web/channel_selector";
+
+Object.assign(Sidebar.components, { ChannelSelector });
