+++ addons/mail/static/tests/discuss_app/discuss_tests.js
@@ -99,159 +99,6 @@ QUnit.test("can change the thread description of #general", async (assert) => {
     assert.verifySteps(["/web/dataset/call_kw/discuss.channel/channel_change_description"]);
 });
 
-QUnit.test("can create a new channel [REQUIRE FOCUS]", async (assert) => {
-    await startServer();
-    const { openDiscuss } = await start({
-        mockRPC(route, params) {
-            if (
-                route.startsWith("/mail") ||
-                route.startsWith("/discuss") ||
-                [
-                    "/web/dataset/call_kw/discuss.channel/search_read",
-                    "/web/dataset/call_kw/discuss.channel/channel_create",
-                ].includes(route)
-            ) {
-                assert.step(route);
-            }
-        },
-    });
-    await openDiscuss();
-    assert.containsNone($, ".o-mail-DiscussCategoryItem");
-
-    await click(".o-mail-DiscussSidebar i[title='Add or join a channel']");
-    await afterNextRender(() => editInput(document.body, ".o-mail-ChannelSelector input", "abc"));
-    await click(".o-mail-ChannelSelector-suggestion");
-    assert.containsOnce($, ".o-mail-DiscussCategoryItem");
-    assert.containsNone($, ".o-mail-Discuss-content .o-mail-Message");
-    assert.verifySteps([
-        "/mail/init_messaging",
-        "/mail/load_message_failures",
-        "/mail/inbox/messages",
-        "/web/dataset/call_kw/discuss.channel/search_read",
-        "/web/dataset/call_kw/discuss.channel/channel_create",
-        "/discuss/channel/messages",
-    ]);
-});
-
-QUnit.test(
-    "do not close channel selector when creating chat conversation after selection",
-    async (assert) => {
-        const pyEnv = await startServer();
-        const partnerId = pyEnv["res.partner"].create({ name: "Mario" });
-        pyEnv["res.users"].create({ partner_id: partnerId });
-        const { openDiscuss } = await start();
-        await openDiscuss();
-        assert.containsNone($, ".o-mail-DiscussCategoryItem");
-
-        await click("i[title='Start a conversation']");
-        await afterNextRender(() =>
-            editInput(document.body, ".o-mail-ChannelSelector input", "mario")
-        );
-        await click(".o-mail-ChannelSelector-suggestion");
-        assert.containsOnce($, ".o-mail-ChannelSelector span[title='Mario']");
-        assert.containsNone($, ".o-mail-DiscussCategoryItem");
-
-        await triggerEvent(document.body, ".o-mail-ChannelSelector input", "keydown", {
-            key: "Backspace",
-        });
-        assert.containsNone($, ".o-mail-ChannelSelector span[title='Mario']");
-
-        await afterNextRender(() =>
-            editInput(document.body, ".o-mail-ChannelSelector input", "mario")
-        );
-        await triggerEvent(document.body, ".o-mail-ChannelSelector input", "keydown", {
-            key: "Enter",
-        });
-        assert.containsOnce($, ".o-mail-ChannelSelector span[title='Mario']");
-        assert.containsNone($, ".o-mail-DiscussCategoryItem");
-    }
-);
-
-QUnit.test("can join a chat conversation", async (assert) => {
-    const pyEnv = await startServer();
-    const partnerId = pyEnv["res.partner"].create({ name: "Mario" });
-    pyEnv["res.users"].create({ partner_id: partnerId });
-    const { openDiscuss } = await start({
-        mockRPC(route, params) {
-            if (
-                route.startsWith("/mail") ||
-                route.startsWith("/discuss") ||
-                ["/web/dataset/call_kw/discuss.channel/channel_get"].includes(route)
-            ) {
-                assert.step(route);
-            }
-            if (route === "/web/dataset/call_kw/discuss.channel/channel_get") {
-                assert.equal(params.kwargs.partners_to[0], partnerId);
-            }
-        },
-    });
-    await openDiscuss();
-    assert.containsNone($, ".o-mail-DiscussCategoryItem");
-
-    await click(".o-mail-DiscussSidebar i[title='Start a conversation']");
-    await afterNextRender(() => editInput(document.body, ".o-mail-ChannelSelector input", "mario"));
-    await click(".o-mail-ChannelSelector-suggestion");
-    await triggerEvent(document.body, ".o-mail-ChannelSelector input", "keydown", {
-        key: "Enter",
-    });
-    assert.containsOnce($, ".o-mail-DiscussCategoryItem");
-    assert.containsNone($, ".o-mail-Message");
-    assert.verifySteps([
-        "/mail/init_messaging",
-        "/mail/load_message_failures",
-        "/mail/inbox/messages",
-        "/web/dataset/call_kw/discuss.channel/channel_get",
-        "/discuss/channel/messages",
-    ]);
-});
-
-QUnit.test("can create a group chat conversation", async (assert) => {
-    const pyEnv = await startServer();
-    const [partnerId_1, partnerId_2] = pyEnv["res.partner"].create([
-        { name: "Mario" },
-        { name: "Luigi" },
-    ]);
-    pyEnv["res.users"].create([{ partner_id: partnerId_1 }, { partner_id: partnerId_2 }]);
-    const { openDiscuss } = await start();
-    await openDiscuss();
-    assert.containsNone($, ".o-mail-DiscussCategoryItem");
-    await click(".o-mail-DiscussSidebar i[title='Start a conversation']");
-    await insertText(".o-mail-ChannelSelector input", "Mario");
-    await click(".o-mail-ChannelSelector-suggestion");
-    await insertText(".o-mail-ChannelSelector input", "Luigi");
-    await click(".o-mail-ChannelSelector-suggestion");
-    await triggerEvent(document.body, ".o-mail-ChannelSelector input", "keydown", {
-        key: "Enter",
-    });
-    assert.containsN($, ".o-mail-DiscussCategoryItem", 1);
-    assert.containsNone($, ".o-mail-Message");
-});
-
-QUnit.test("chat search should display no result when no matches found", async (assert) => {
-    const { openDiscuss } = await start();
-    await openDiscuss();
-    await click(".o-mail-DiscussSidebar i[title='Start a conversation']");
-    await insertText(".o-mail-ChannelSelector", "Rainbow Panda");
-    assert.containsOnce($, ".o-mail-ChannelSelector-suggestion:contains(No results found)");
-});
-
-QUnit.test(
-    "chat search should not be visible when clicking outside of the field",
-    async (assert) => {
-        const pyEnv = await startServer();
-        const partnerId = pyEnv["res.partner"].create({ name: "Panda" });
-        pyEnv["res.users"].create({ partner_id: partnerId });
-        const { openDiscuss } = await start();
-        await openDiscuss();
-        assert.containsNone($, ".o-mail-DiscussCategoryItem");
-        await click(".o-mail-DiscussSidebar i[title='Start a conversation']");
-        await insertText(".o-mail-ChannelSelector", "Panda");
-        assert.containsOnce($, ".o-mail-ChannelSelector-suggestion");
-        await click(".o-mail-DiscussSidebar");
-        assert.containsNone($, ".o-mail-ChannelSelector-suggestion");
-    }
-);
-
 QUnit.test("Message following a notification should not be squashed", async (assert) => {
     const pyEnv = await startServer();
     const channelId = pyEnv["discuss.channel"].create({
@@ -657,20 +504,6 @@ QUnit.test("sidebar: change active", async (assert) => {
     await click("button:contains(Starred)");
     assert.doesNotHaveClass($("button:contains(Inbox)"), "o-active");
     assert.hasClass($("button:contains(Starred)"), "o-active");
-});
-
-QUnit.test("sidebar: add channel", async (assert) => {
-    const { openDiscuss } = await start();
-    await openDiscuss();
-    assert.containsOnce($, ".o-mail-DiscussCategory-channel .o-mail-DiscussCategory-add");
-    assert.hasAttrValue(
-        $(".o-mail-DiscussCategory-channel .o-mail-DiscussCategory-add")[0],
-        "title",
-        "Add or join a channel"
-    );
-    await click(".o-mail-DiscussCategory-channel .o-mail-DiscussCategory-add");
-    assert.containsOnce($, ".o-mail-ChannelSelector");
-    assert.containsOnce($, ".o-mail-ChannelSelector input[placeholder='Add or join a channel']");
 });
 
 QUnit.test("sidebar: basic channel rendering", async (assert) => {
@@ -1488,22 +1321,6 @@ QUnit.test("Channel is added to discuss after invitation", async (assert) => {
     });
     assert.containsOnce($, ".o-mail-DiscussCategoryItem:contains(General)");
     assert.verifySteps(["You have been invited to #General"]);
-});
-
-QUnit.test("Chat is added to discuss on other tab that the one that joined", async (assert) => {
-    const pyEnv = await startServer();
-    const partnerId = pyEnv["res.partner"].create({ name: "Jerry Golay" });
-    pyEnv["res.users"].create({ partner_id: partnerId });
-    const tab1 = await start({ asTab: true });
-    const tab2 = await start({ asTab: true });
-    await tab1.openDiscuss();
-    await tab2.openDiscuss();
-    await tab1.click(".o-mail-DiscussCategory-chat .o-mail-DiscussCategory-add");
-    await tab1.insertText(".o-mail-ChannelSelector input", "Jer");
-    await tab1.click(".o-mail-ChannelSelector-suggestion");
-    await afterNextRender(() => triggerHotkey("Enter"));
-    assert.containsOnce(tab1.target, ".o-mail-DiscussCategoryItem:contains(Jerry Golay)");
-    assert.containsOnce(tab2.target, ".o-mail-DiscussCategoryItem:contains(Jerry Golay)");
 });
 
 QUnit.test("select another mailbox", async (assert) => {
