+++ addons/mail/static/tests/discuss_app/sidebar_tests.js
@@ -5,7 +5,6 @@ import {
     afterNextRender,
     click,
     insertText,
-    nextAnimationFrame,
     start,
     startServer,
     waitUntil,
@@ -13,7 +12,7 @@ import {
 
 import { getOrigin } from "@web/core/utils/urls";
 import { makeFakeNotificationService } from "@web/../tests/helpers/mock_services";
-import { editInput, makeDeferred, nextTick } from "@web/../tests/helpers/utils";
+import { editInput, nextTick } from "@web/../tests/helpers/utils";
 
 QUnit.module("discuss sidebar");
 
@@ -631,65 +630,6 @@ QUnit.test("chat - states: close manually by clicking the title", async (assert)
     await click(".o-mail-DiscussCategory:contains(Direct messages) div");
     assert.containsNone($, ".o-mail-DiscussCategoryItem");
 });
-
-QUnit.test("sidebar find shows channels matching search term", async (assert) => {
-    const pyEnv = await startServer();
-    pyEnv["discuss.channel"].create({
-        channel_member_ids: [],
-        channel_type: "channel",
-        group_public_id: false,
-        name: "test",
-    });
-    const def = makeDeferred();
-    const { openDiscuss } = await start({
-        async mockRPC(route, args) {
-            if (args.method === "search_read") {
-                def.resolve();
-            }
-        },
-    });
-    await openDiscuss();
-    await click(".o-mail-DiscussCategory-add");
-    await insertText(".o-mail-ChannelSelector input", "test");
-    await def;
-    await nextAnimationFrame(); // ensures search_read rpc is rendered.
-    // When searching for a single existing channel, the results list will have at least 2 lines:
-    // One for the existing channel itself
-    // One for creating a channel with the search term
-    assert.containsN($, ".o-mail-NavigableList-item", 2);
-    assert.containsN($, ".o-mail-NavigableList-item:contains(test)", 2);
-});
-
-QUnit.test(
-    "sidebar find shows channels matching search term even when user is member",
-    async (assert) => {
-        const pyEnv = await startServer();
-        pyEnv["discuss.channel"].create({
-            channel_member_ids: [Command.create({ partner_id: pyEnv.currentPartnerId })],
-            channel_type: "channel",
-            group_public_id: false,
-            name: "test",
-        });
-        const def = makeDeferred();
-        const { openDiscuss } = await start({
-            async mockRPC(route, args) {
-                if (args.method === "search_read") {
-                    def.resolve();
-                }
-            },
-        });
-        await openDiscuss();
-        await click(".o-mail-DiscussCategory-add");
-        await insertText(".o-mail-ChannelSelector input", "test");
-        await def;
-        await nextAnimationFrame(); // ensures search_read rpc is rendered.
-        // When searching for a single existing channel, the results list will have at least 2 lines:
-        // One for the existing channel itself
-        // One for creating a channel with the search term
-        assert.containsN($, ".o-mail-NavigableList-item", 2);
-        assert.containsN($, ".o-mail-NavigableList-item:contains(test)", 2);
-    }
-);
 
 QUnit.test("sidebar channels should be ordered case insensitive alphabetically", async (assert) => {
     const pyEnv = await startServer();
